Add delete account functionality to settings page
In this update, the delete account functionality was added to the user settings page. It involved importing new modules and defining the action function for deleting an account. An appropriate response is returned if invalid action is detected. In addition, a schema for the delete account form was created and referenced in the related components.