Refactor i18n module and enhance error handling in admin features
In this commit, the i18n module was refactored to move the parseAcceptLanguageHeader function to its own file for better modularity. The i18n server file is also deleted and recreated to better handle case sensitivity in various operating systems. Additionally, error handling in the admin features of the application was enhanced to provide better logs in the console for troubleshooting. The UI for admin routes was updated to better protect certain routes and to enhance user experience.