Add React and necessary configurations to WordPress and Keystatic packages
Several changes have been made to the WordPress and Keystatic packages. React has been added as a dependency in both packages. Subsequently, the tsconfig.json file had to be modified to include the JSX module. Additional updates have also been made to ensure compatibility with the newly added module. The types of certain objects had to be changed to work with the new module.