Added dotenv-cli and start script with environment variables
The dotenv-cli package has been added to the project to load environment variables from the .env file. Also, two new scripts `start:with-env` and `with-env` have been added to the package.json file to support running the project with these environment variables. Lastly, in the Wordpress client, the import statement of the Cms and CmsClient was changed to import types only.