Update user settings and refactor components for improved code readability
Updated the PersonalAccountSettingsPage to pass the userId prop directly to its child components. The useSignUp hook now invalidates all queries on sign out. A new React Query Provider component was created and QueryClientProvider was replaced with it in root-providers component for improved code maintenance and readability. Refactored several components to pass the userId prop directly, and updated hooks to use the new userId prop. Added cache invalidation on sign in/sign out and cleaning the cache on server side.