Refactor subscription payload building across services (#8)
* Refactor subscription payload building across services

Extracted the logic of building the subscription payload into individual services. This allows for more maintainable and isolated code, making it easier to update or modify the process in the future. Changes occurred in both the Stripe and Lemon Squeezy services. Also added the ability to retrieve a subscription from the provider.

* Remove database query for account ID in Lemon-Squeezy service

The modification in the codebase removes the need for querying the database to obtain an account ID in the Lemon-Squeezy billing-strategy service. This change was motivated by the realization that the Lemon Squeezy API does not guarantee the presence of an account ID. Hence

* Refactor billing service and enhance auth callback service

The billing strategy provider service's getSubscription method is refactored for code readability. In the Supabase auth callback service, handling for the 'next' parameter is added. If a 'next' path exists, it is now appended to the search parameters, improving the flow control on user redirection.