Add version updater feature
Implemented a new auxiliary feature to periodically check for new app versions and alert the user when a new version is available. This scheme improves user experience by ensuring users are always running the most updated version of the application. The version updater is enabled by default but can be managed using a feature flag.