Add terms and conditions checkbox to sign-up forms (#10)
A new checkbox has been added to the user sign-up forms for both 'Email Password' and 'Magic link' sign-up methods. It asks users to accept the terms and conditions as well as privacy policy before signing up. The checkbox is visible or hidden based on the 'displayTermsCheckbox' boolean value in the auth configuration.