Add Marketing module and related components
Added an extensive set of components within a new Marketing module. This includes UI elements such as headers, footers, call-to-action buttons, cards, and newsletter signup functionality. The corresponding files were added and exported in the index.tsx file of the module. The addition of the Marketing module was also accounted for, in the package.json file.